Fix: remove redundant buffer clearing in main loop and update EEPROM function parameter type

+++ ch32v003fun/driver.h
@@ -16,11 +16,11 @@ void MY_I2C_init(){
 }
 #define EEPROM_ADDR 0x52 // obtained from i2c_scan(), before shifting by 1 bit
 
-i2c_result_e EEPROM_write(uint16_t regAddr, uint8_t *data, uint8_t sz){
+i2c_result_e EEPROM_write(uint16_t regAddr, uint8_t *data, uint16_t sz){
     return i2c_write(EEPROM_ADDR, regAddr, I2C_REGADDR_2B, data, sz);
 }
 
-i2c_result_e EEPROM_read(uint16_t regAddr, uint8_t *data, uint8_t sz){
+i2c_result_e EEPROM_read(uint16_t regAddr, uint8_t *data, uint16_t sz){
     return i2c_read(EEPROM_ADDR, regAddr, I2C_REGADDR_2B, data, sz);
 }
